requirements.txt
asgiref==3.10.0
contourpy==1.3.2
cycler==0.12.1
distlib==0.4.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
fonttools==4.58.0
graphviz==0.20.3
kiwisolver==1.4.8
matplotlib==3.10.3
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
platformdirs==4.5.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.35.3
